requirements.txt
asgiref==3.8.1
async-timeout==5.0.0
certifi==2024.8.30
Django==5.1.2
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.11.2
django-nine==0.2.7
django-redis==5.4.0
djangorestframework==3.15.2
elastic-transport==8.15.1
elasticsearch==8.15.1
elasticsearch-dsl==8.15.4
packaging==24.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
